websiteScanner/requirements.txt
asgiref==3.6.0
blinker==1.4
certifi==2022.12.7
charset-normalizer==3.0.1
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
djangorestframework==3.14.0
drf-yasg==1.21.4
filelock==3.6.0
future==0.18.3
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.0
platformdirs==2.5.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0
python-whois==0.8.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.3
systemd-python==234
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.13.0+ds
wadllib==1.3.6
zipp==1.0.0
